requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
